feat: wire SqlNormalizer into Validator.validate()
